Combine tests into common-http-errors.js
* Add `woql` to these.
